Refactor: src layout, clean packaging, remove version module
