+++ .gitignore
@@ -2,6 +2,7 @@
 __pycache__/
 *.pyc
 *.pyo
+*.egg-info/
 .ruff_cache/
 
 # Virtual environments
