+++ pyproject.toml
@@ -2,6 +2,9 @@
 name = "tapmap"
 version = "1.5.0"
 
+[project.scripts]
+tapmap = "tapmap.app:main"
+
 [tool.ruff]
 line-length = 100
 target-version = "py310"
@@ -21,5 +24,11 @@ per-file-ignores = { "tests/*" = ["D101", "D102", "D103"] }
 [tool.ruff.lint.pydocstyle]
 convention = "google"
 
+[tool.setuptools]
+package-dir = {"" = "src"}
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
 [tool.pytest.ini_options]
-pythonpath = ["."]
+pythonpath = ["src"]
