+++ src/tapmap/__init__.py
@@ -1,0 +1,6 @@
+"""TapMap application package.
+
+Exposes application entry point and package version.
+"""
+
+__version__ = "1.5.0"
