+++ src/tapmap/model/netinfo.py
