+++ src/tapmap/model/netinfo_lsof.py
