+++ src/tapmap/state/insights.py
