+++ src/tapmap/state/status_cache.py
