+++ src/tapmap/ui/insights_view.py
